# Introduced vectorized batch norm backward kernels