--- a/src/ATen/native/xpu/sycl/BatchNormKernels.cpp
+++ b/src/ATen/native/xpu/sycl/BatchNormKernels.cpp
@@ -2868,6 +2868,144 @@ struct BatchNormBackwardElemtChannelsLastKernelFunctor {
   const int64_t world_size_;
 };
 
+template <
+    int PARALLEL_LOADS,
+    int VEC_SIZE,
+    typename scalar_t,
+    typename accscalar_t,
+    typename layerscalar_t,
+    bool USE_COUNTS = false>
+struct BatchNormBackwardElemtChannelsLastVectorizedKernelFunctor {
+  void operator()(sycl::nd_item<2> item) const {
+    accscalar_t norm_fct;
+    if constexpr (USE_COUNTS) {
+      int64_t total_numel = 0;
+      for (int i = 0; i < world_size_; i++) {
+        total_numel += numel_[i];
+      }
+      norm_fct =
+          static_cast<accscalar_t>(1) / static_cast<accscalar_t>(total_numel);
+    } else {
+      norm_fct = norm_fct_;
+    }
+
+    // tensor dimension (m,c)
+    // loop along m dimension
+    int inner_loop_stride = item.get_local_range(0) * item.get_group_range(0);
+
+    // offset along m dimension
+    int m_offset = item.get_global_id(0);
+    int c_vec_offset = item.get_global_id(1) * VEC_SIZE;
+
+    if (c_vec_offset >= stride_ || m_offset >= reduction_size_) {
+      return;
+    }
+
+    int loop_count =
+        1 + (reduction_size_ - 1) / (inner_loop_stride * PARALLEL_LOADS);
+    int address_base = m_offset * stride_ + c_vec_offset;
+    int address_increment = inner_loop_stride * stride_;
+
+    for (int i = 0; i < loop_count; i++) {
+#pragma unroll
+      for (int j = 0; j < PARALLEL_LOADS; j++) {
+        if (m_offset < reduction_size_) {
+          using vec_t = memory::aligned_vector<scalar_t, VEC_SIZE>;
+          auto grad_output_vec = *reinterpret_cast<vec_t*>(
+              const_cast<scalar_t*>(&grad_output_[address_base]));
+          auto input_vec = *reinterpret_cast<vec_t*>(
+              const_cast<scalar_t*>(&input_[address_base]));
+          vec_t grad_input_vec;
+
+#pragma unroll
+          for (int v = 0; v < VEC_SIZE; ++v) {
+            int c_offset = c_vec_offset + v;
+
+            auto m_c = mean_[c_offset];
+            auto m_dy_c = sum_dy_[c_offset] * norm_fct;
+            auto factor_1_c = inv_std_[c_offset];
+            auto factor_2_c =
+                (weight_ == nullptr
+                     ? accscalar_t(1.0)
+                     : static_cast<accscalar_t>(weight_[c_offset])) *
+                factor_1_c;
+            factor_1_c =
+                factor_1_c * factor_1_c * sum_dy_xmu_[c_offset] * norm_fct;
+
+            grad_input_vec[v] = static_cast<scalar_t>(
+                (static_cast<accscalar_t>(grad_output_vec[v]) - m_dy_c -
+                 (static_cast<accscalar_t>(input_vec[v]) - m_c) * factor_1_c) *
+                factor_2_c);
+          }
+
+          *reinterpret_cast<vec_t*>(&grad_input_[address_base]) =
+              grad_input_vec;
+        }
+        m_offset += inner_loop_stride;
+        address_base += address_increment;
+      }
+    }
+  }
+
+  BatchNormBackwardElemtChannelsLastVectorizedKernelFunctor(
+      const scalar_t* RESTRICT grad_output,
+      const scalar_t* RESTRICT input,
+      const accscalar_t* RESTRICT mean,
+      const accscalar_t* RESTRICT inv_std,
+      const layerscalar_t* RESTRICT weight,
+      const accscalar_t* RESTRICT sum_dy,
+      const accscalar_t* RESTRICT sum_dy_xmu,
+      scalar_t* RESTRICT grad_input,
+      const accscalar_t norm_fct,
+      const int reduction_size,
+      const int stride,
+      const int* RESTRICT numel = nullptr,
+      const int64_t world_size = 0)
+      : grad_output_(grad_output),
+        input_(input),
+        mean_(mean),
+        inv_std_(inv_std),
+        weight_(weight),
+        sum_dy_(sum_dy),
+        sum_dy_xmu_(sum_dy_xmu),
+        grad_input_(grad_input),
+        norm_fct_(norm_fct),
+        reduction_size_(reduction_size),
+        stride_(stride),
+        numel_(numel),
+        world_size_(world_size) {}
+
+ private:
+  const scalar_t* RESTRICT grad_output_;
+  const scalar_t* RESTRICT input_;
+  const accscalar_t* RESTRICT mean_;
+  const accscalar_t* RESTRICT inv_std_;
+  const layerscalar_t* RESTRICT weight_;
+  const accscalar_t* RESTRICT sum_dy_;
+  const accscalar_t* RESTRICT sum_dy_xmu_;
+  scalar_t* RESTRICT grad_input_;
+  const accscalar_t norm_fct_;
+  const int reduction_size_;
+  const int stride_;
+  const int* RESTRICT numel_;
+  const int64_t world_size_;
+};
+
+template <typename scalar_t, int VEC_SIZE>
+bool can_use_batch_norm_bwd_cnl_vec_kernel(
+    const scalar_t* grad_output_ptr,
+    const scalar_t* input_ptr,
+    scalar_t* grad_input_ptr,
+    int stride) {
+  return (
+      memory::can_vectorize_up_to<scalar_t>((char*)grad_output_ptr) >=
+          VEC_SIZE &&
+      memory::can_vectorize_up_to<scalar_t>((char*)input_ptr) >= VEC_SIZE &&
+      memory::can_vectorize_up_to<scalar_t>((char*)grad_input_ptr) >=
+          VEC_SIZE &&
+      stride % VEC_SIZE == 0);
+}
+
 at::Tensor batch_norm_backward_elemt_channels_last_template(
     const at::Tensor& grad_output,
     const at::Tensor& input,
@@ -2882,11 +3020,6 @@ at::Tensor batch_norm_backward_elemt_channels_last_template(
 
   // Input is guarunteed to be channels-last compatible
   at::Tensor grad_input = at::empty_like(input);
-
-  auto config = get_adaptive_launch_config(
-      reduction_size, stride, true, ELEMENTS_PER_WORK_ITEM);
-  auto global_range = std::get<0>(config);
-  auto local_range = std::get<1>(config);
   auto& queue = getCurrentSYCLQueue();
 
   AT_DISPATCH_FLOATING_TYPES_AND2(
@@ -2896,46 +3029,103 @@ at::Tensor batch_norm_backward_elemt_channels_last_template(
       "batchnorm_backward_element_xpu",
       [&] {
         using accscalar_t = at::acc_type_device<scalar_t, kXPU>;
+        constexpr int VEC_SIZE = 2;
+        auto grad_output_ptr = grad_output.const_data_ptr<scalar_t>();
+        auto input_ptr = input.const_data_ptr<scalar_t>();
+        auto grad_input_ptr = grad_input.mutable_data_ptr<scalar_t>();
 
-        if (weight.defined() && weight.scalar_type() != input.scalar_type()) {
-          auto kfn = BatchNormBackwardElemtChannelsLastKernelFunctor<
-              ELEMENTS_PER_ITER,
-              scalar_t,
-              accscalar_t,
-              accscalar_t>(
-              grad_output.const_data_ptr<scalar_t>(),
-              input.const_data_ptr<scalar_t>(),
-              mean.const_data_ptr<accscalar_t>(),
-              inv_std.const_data_ptr<accscalar_t>(),
-              weight.const_data_ptr<accscalar_t>(),
-              sum_dy.const_data_ptr<accscalar_t>(),
-              sum_dy_xmu.const_data_ptr<accscalar_t>(),
-              grad_input.mutable_data_ptr<scalar_t>(),
-              static_cast<accscalar_t>(norm_fct),
-              reduction_size,
-              stride);
-          sycl_kernel_submit(global_range, local_range, queue, kfn);
+        if (can_use_batch_norm_bwd_cnl_vec_kernel<scalar_t, VEC_SIZE>(
+                grad_output_ptr, input_ptr, grad_input_ptr, stride)) {
+          auto config = get_adaptive_launch_config(
+              reduction_size, stride / VEC_SIZE, true, ELEMENTS_PER_WORK_ITEM);
+          auto global_range = std::get<0>(config);
+          auto local_range = std::get<1>(config);
+          if (weight.defined() && weight.scalar_type() != input.scalar_type()) {
+            auto kfn =
+                BatchNormBackwardElemtChannelsLastVectorizedKernelFunctor<
+                    ELEMENTS_PER_ITER,
+                    VEC_SIZE,
+                    scalar_t,
+                    accscalar_t,
+                    accscalar_t>(
+                    grad_output_ptr,
+                    input_ptr,
+                    mean.const_data_ptr<accscalar_t>(),
+                    inv_std.const_data_ptr<accscalar_t>(),
+                    weight.const_data_ptr<accscalar_t>(),
+                    sum_dy.const_data_ptr<accscalar_t>(),
+                    sum_dy_xmu.const_data_ptr<accscalar_t>(),
+                    grad_input_ptr,
+                    static_cast<accscalar_t>(norm_fct),
+                    reduction_size,
+                    stride);
+            sycl_kernel_submit(global_range, local_range, queue, kfn);
+          } else {
+            auto kfn =
+                BatchNormBackwardElemtChannelsLastVectorizedKernelFunctor<
+                    ELEMENTS_PER_ITER,
+                    VEC_SIZE,
+                    scalar_t,
+                    accscalar_t,
+                    scalar_t>(
+                    grad_output_ptr,
+                    input_ptr,
+                    mean.const_data_ptr<accscalar_t>(),
+                    inv_std.const_data_ptr<accscalar_t>(),
+                    weight.defined() ? weight.const_data_ptr<scalar_t>()
+                                     : nullptr,
+                    sum_dy.const_data_ptr<accscalar_t>(),
+                    sum_dy_xmu.const_data_ptr<accscalar_t>(),
+                    grad_input_ptr,
+                    static_cast<accscalar_t>(norm_fct),
+                    reduction_size,
+                    stride);
+            sycl_kernel_submit(global_range, local_range, queue, kfn);
+          }
         } else {
-          auto kfn = BatchNormBackwardElemtChannelsLastKernelFunctor<
-              ELEMENTS_PER_ITER,
-              scalar_t,
-              accscalar_t,
-              scalar_t>(
-              grad_output.const_data_ptr<scalar_t>(),
-              input.const_data_ptr<scalar_t>(),
-              mean.const_data_ptr<accscalar_t>(),
-              inv_std.const_data_ptr<accscalar_t>(),
-              weight.defined() ? weight.const_data_ptr<scalar_t>() : nullptr,
-              sum_dy.const_data_ptr<accscalar_t>(),
-              sum_dy_xmu.const_data_ptr<accscalar_t>(),
-              grad_input.mutable_data_ptr<scalar_t>(),
-              static_cast<accscalar_t>(norm_fct),
-              reduction_size,
-              stride);
-          sycl_kernel_submit(global_range, local_range, queue, kfn);
+          auto config = get_adaptive_launch_config(
+              reduction_size, stride, true, ELEMENTS_PER_WORK_ITEM);
+          auto global_range = std::get<0>(config);
+          auto local_range = std::get<1>(config);
+          if (weight.defined() && weight.scalar_type() != input.scalar_type()) {
+            auto kfn = BatchNormBackwardElemtChannelsLastKernelFunctor<
+                ELEMENTS_PER_ITER,
+                scalar_t,
+                accscalar_t,
+                accscalar_t>(
+                grad_output_ptr,
+                input_ptr,
+                mean.const_data_ptr<accscalar_t>(),
+                inv_std.const_data_ptr<accscalar_t>(),
+                weight.const_data_ptr<accscalar_t>(),
+                sum_dy.const_data_ptr<accscalar_t>(),
+                sum_dy_xmu.const_data_ptr<accscalar_t>(),
+                grad_input_ptr,
+                static_cast<accscalar_t>(norm_fct),
+                reduction_size,
+                stride);
+            sycl_kernel_submit(global_range, local_range, queue, kfn);
+          } else {
+            auto kfn = BatchNormBackwardElemtChannelsLastKernelFunctor<
+                ELEMENTS_PER_ITER,
+                scalar_t,
+                accscalar_t,
+                scalar_t>(
+                grad_output_ptr,
+                input_ptr,
+                mean.const_data_ptr<accscalar_t>(),
+                inv_std.const_data_ptr<accscalar_t>(),
+                weight.defined() ? weight.const_data_ptr<scalar_t>() : nullptr,
+                sum_dy.const_data_ptr<accscalar_t>(),
+                sum_dy_xmu.const_data_ptr<accscalar_t>(),
+                grad_input_ptr,
+                static_cast<accscalar_t>(norm_fct),
+                reduction_size,
+                stride);
+            sycl_kernel_submit(global_range, local_range, queue, kfn);
+          }
         }
       });
-
   return grad_input;
 }
 
@@ -2953,12 +3143,8 @@ at::Tensor batch_norm_backward_elemt_channels_last_template(
 
   // Input is guarunteed to be channels-last compatible
   at::Tensor grad_input = at::empty_like(input);
-
-  auto config = get_adaptive_launch_config(
-      reduction_size, stride, false, ELEMENTS_PER_WORK_ITEM);
-  auto global_range = std::get<0>(config);
-  auto local_range = std::get<1>(config);
   auto& queue = getCurrentSYCLQueue();
+  constexpr int VEC_SIZE = 2;
 
   if (weight.defined() && weight.scalar_type() != input.scalar_type()) {
     AT_DISPATCH_FLOATING_TYPES_AND2(
@@ -2968,26 +3154,64 @@ at::Tensor batch_norm_backward_elemt_channels_last_template(
         "batchnorm_backward_element_xpu",
         [&] {
           using accscalar_t = acc_type_device<scalar_t, kXPU>;
-          auto kfn = BatchNormBackwardElemtChannelsLastKernelFunctor<
-              ELEMENTS_PER_ITER,
-              scalar_t,
-              accscalar_t,
-              accscalar_t,
-              true>(
-              grad_output.const_data_ptr<scalar_t>(),
-              input.const_data_ptr<scalar_t>(),
-              mean.const_data_ptr<accscalar_t>(),
-              inv_std.const_data_ptr<accscalar_t>(),
-              weight.const_data_ptr<accscalar_t>(),
-              sum_dy.const_data_ptr<accscalar_t>(),
-              sum_dy_xmu.const_data_ptr<accscalar_t>(),
-              grad_input.mutable_data_ptr<scalar_t>(),
-              0,
-              reduction_size,
-              stride,
-              count.const_data_ptr<int>(),
-              count.numel());
-          sycl_kernel_submit(global_range, local_range, queue, kfn);
+          auto grad_output_ptr = grad_output.const_data_ptr<scalar_t>();
+          auto input_ptr = input.const_data_ptr<scalar_t>();
+          auto grad_input_ptr = grad_input.mutable_data_ptr<scalar_t>();
+
+          if (!can_use_batch_norm_bwd_cnl_vec_kernel<scalar_t, VEC_SIZE>(
+                  grad_output_ptr, input_ptr, grad_input_ptr, stride)) {
+            auto config = get_adaptive_launch_config(
+                reduction_size, stride, false, ELEMENTS_PER_WORK_ITEM);
+            auto global_range = std::get<0>(config);
+            auto local_range = std::get<1>(config);
+            auto kfn = BatchNormBackwardElemtChannelsLastKernelFunctor<
+                ELEMENTS_PER_ITER,
+                scalar_t,
+                accscalar_t,
+                accscalar_t,
+                true>(
+                grad_output_ptr,
+                input_ptr,
+                mean.const_data_ptr<accscalar_t>(),
+                inv_std.const_data_ptr<accscalar_t>(),
+                weight.const_data_ptr<accscalar_t>(),
+                sum_dy.const_data_ptr<accscalar_t>(),
+                sum_dy_xmu.const_data_ptr<accscalar_t>(),
+                grad_input_ptr,
+                0,
+                reduction_size,
+                stride,
+                count.const_data_ptr<int>(),
+                count.numel());
+            sycl_kernel_submit(global_range, local_range, queue, kfn);
+          } else {
+            auto config = get_adaptive_launch_config(
+                reduction_size, stride / VEC_SIZE, false, ELEMENTS_PER_WORK_ITEM);
+            auto global_range = std::get<0>(config);
+            auto local_range = std::get<1>(config);
+            auto kfn =
+                BatchNormBackwardElemtChannelsLastVectorizedKernelFunctor<
+                    ELEMENTS_PER_ITER,
+                    VEC_SIZE,
+                    scalar_t,
+                    accscalar_t,
+                    accscalar_t,
+                    true>(
+                    grad_output_ptr,
+                    input_ptr,
+                    mean.const_data_ptr<accscalar_t>(),
+                    inv_std.const_data_ptr<accscalar_t>(),
+                    weight.const_data_ptr<accscalar_t>(),
+                    sum_dy.const_data_ptr<accscalar_t>(),
+                    sum_dy_xmu.const_data_ptr<accscalar_t>(),
+                    grad_input_ptr,
+                    0,
+                    reduction_size,
+                    stride,
+                    count.const_data_ptr<int>(),
+                    count.numel());
+            sycl_kernel_submit(global_range, local_range, queue, kfn);
+          }
         });
   } else {
     if (weight.defined()) {
@@ -3005,26 +3229,65 @@ at::Tensor batch_norm_backward_elemt_channels_last_template(
         "batchnorm_backward_element_xpu",
         [&] {
           using accscalar_t = acc_type_device<scalar_t, kXPU>;
-          auto kfn = BatchNormBackwardElemtChannelsLastKernelFunctor<
-              ELEMENTS_PER_ITER,
-              scalar_t,
-              accscalar_t,
-              scalar_t,
-              true>(
-              grad_output.const_data_ptr<scalar_t>(),
-              input.const_data_ptr<scalar_t>(),
-              mean.const_data_ptr<accscalar_t>(),
-              inv_std.const_data_ptr<accscalar_t>(),
-              weight.defined() ? weight.const_data_ptr<scalar_t>() : nullptr,
-              sum_dy.const_data_ptr<accscalar_t>(),
-              sum_dy_xmu.const_data_ptr<accscalar_t>(),
-              grad_input.mutable_data_ptr<scalar_t>(),
-              0,
-              reduction_size,
-              stride,
-              count.const_data_ptr<int>(),
-              count.numel());
-          sycl_kernel_submit(global_range, local_range, queue, kfn);
+          auto grad_output_ptr = grad_output.const_data_ptr<scalar_t>();
+          auto input_ptr = input.const_data_ptr<scalar_t>();
+          auto grad_input_ptr = grad_input.mutable_data_ptr<scalar_t>();
+
+          if (!can_use_batch_norm_bwd_cnl_vec_kernel<scalar_t, VEC_SIZE>(
+                  grad_output_ptr, input_ptr, grad_input_ptr, stride)) {
+            auto config = get_adaptive_launch_config(
+                reduction_size, stride, false, ELEMENTS_PER_WORK_ITEM);
+            auto global_range = std::get<0>(config);
+            auto local_range = std::get<1>(config);
+            auto kfn = BatchNormBackwardElemtChannelsLastKernelFunctor<
+                ELEMENTS_PER_ITER,
+                scalar_t,
+                accscalar_t,
+                scalar_t,
+                true>(
+                grad_output_ptr,
+                input_ptr,
+                mean.const_data_ptr<accscalar_t>(),
+                inv_std.const_data_ptr<accscalar_t>(),
+                weight.defined() ? weight.const_data_ptr<scalar_t>() : nullptr,
+                sum_dy.const_data_ptr<accscalar_t>(),
+                sum_dy_xmu.const_data_ptr<accscalar_t>(),
+                grad_input_ptr,
+                0,
+                reduction_size,
+                stride,
+                count.const_data_ptr<int>(),
+                count.numel());
+            sycl_kernel_submit(global_range, local_range, queue, kfn);
+          } else {
+            auto config = get_adaptive_launch_config(
+                reduction_size, stride / 2, false, ELEMENTS_PER_WORK_ITEM);
+            auto global_range = std::get<0>(config);
+            auto local_range = std::get<1>(config);
+            auto kfn =
+                BatchNormBackwardElemtChannelsLastVectorizedKernelFunctor<
+                    ELEMENTS_PER_ITER,
+                    VEC_SIZE,
+                    scalar_t,
+                    accscalar_t,
+                    scalar_t,
+                    true>(
+                    grad_output_ptr,
+                    input_ptr,
+                    mean.const_data_ptr<accscalar_t>(),
+                    inv_std.const_data_ptr<accscalar_t>(),
+                    weight.defined() ? weight.const_data_ptr<scalar_t>()
+                                     : nullptr,
+                    sum_dy.const_data_ptr<accscalar_t>(),
+                    sum_dy_xmu.const_data_ptr<accscalar_t>(),
+                    grad_input_ptr,
+                    0,
+                    reduction_size,
+                    stride,
+                    count.const_data_ptr<int>(),
+                    count.numel());
+            sycl_kernel_submit(global_range, local_range, queue, kfn);
+          }
         });
   }