Add vectorization implementations of batch norm backward kernels, with speeding up by up to 1.27x for channels-last cases and 1.32x for normal cases.